Shuffle control board internals to be more generic
Create auv_control_board to share common protocol between robot control
board and MEB, which uses the same communication spec.
