+++ Cargo.toml
@@ -26,6 +26,7 @@ itertools = "0.11.0" # Enhance iterators
 num-traits = "0.2.16" # Numeric generics
 derive-getters = "0.3.0" # Getter macro
 async-trait = "0.1.73" # Async fns in trait definitions
+futures = "0.3.28" # Futures utilities
 
 [dev-dependencies]
 assert_approx_eq = "1.1.0" # Floating point eq
