+++ src/comms/mod.rs
@@ -1,1 +1,2 @@
+pub mod auv_control_board;
 pub mod control_board;
